// apps/oas-core-validator-demo/src/generated/openapi.types.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export type paths = Record<string, never>;
export type webhooks = Record<string, never>;
export interface components {
  schemas: {
    /** @enum {string} */
    Tier: "free" | "pro";
    User: {
      id: string;
      /** Format: email */
      email: string;
      tier: components["schemas"]["Tier"];
      isVerified: boolean;
      /** @default [] */
      tags: string[];
    };
    VerifiedUser: components["schemas"]["User"] & {
      /** @constant */
      isVerified: true;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}
export type $defs = Record<string, never>;
export type operations = Record<string, never>;